+++ CHANGELOG.md
@@ -6,8 +6,8 @@ The formatter is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
-- BREAKING CHANGE: Add new JavaScript implementation ([#4](https://github.com/cucumber/pretty-formatter/pull/4))
+### Changed
+- Added JavaScript implementation, replacing [cucumber-js-pretty-formatter](https://github.com/cucumber/cucumber-js-pretty-formatter) ([#4](https://github.com/cucumber/pretty-formatter/pull/4))
 
 ## [0.3.0] - 2025-07-10
 ### Added