+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The formatter is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add attachments option ([#19](https://github.com/cucumber/pretty-formatter/pull/19))
 
 ## [2.0.1] - 2025-07-19