+++ CHANGELOG.md
@@ -6,11 +6,11 @@ The formatter is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
+### Added
 - Add attachments option ([#19](https://github.com/cucumber/pretty-formatter/pull/19))
 
 ### Fixed
-- Print stacktrace with message and type ([#14](https://github.com/cucumber/pretty-formatter/pull/14))
+- [Javascript] Print stacktrace with message and type ([#14](https://github.com/cucumber/pretty-formatter/pull/14))
 
 ## [2.0.1] - 2025-07-19
 ### Fixed