+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add attachments option ([#19](https://github.com/cucumber/pretty-formatter/pull/19))
 
 ### Fixed
-- [Javascript] Print stacktrace with message and type ([#14](https://github.com/cucumber/pretty-formatter/pull/14))
+- [JavaScript] Print stacktrace with message and type ([#14](https://github.com/cucumber/pretty-formatter/pull/14))
 
 ## [2.0.1] - 2025-07-19
 ### Fixed