+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The formatter is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.1.0] - 2025-08-17
 ### Added
 - Add attachments option ([#19](https://github.com/cucumber/pretty-formatter/pull/19))
 
@@ -35,7 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Java implementation ([#1](https://github.com/cucumber/pretty-formatter/pull/1) M.P. Korstanje)
 
-[Unreleased]: https://github.com/cucumber/pretty-formatter/compare/v2.0.1...HEAD
+[Unreleased]: https://github.com/cucumber/pretty-formatter/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/cucumber/pretty-formatter/compare/v2.0.1...v2.1.0
 [2.0.1]: https://github.com/cucumber/pretty-formatter/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/cucumber/pretty-formatter/compare/v0.3.0...v2.0.0
 [0.3.0]: https://github.com/cucumber/pretty-formatter/compare/v0.2.0...v0.3.0